working on send_otp and create viewTest

+++ account/views.py
@@ -1,14 +1,16 @@
 from django.contrib import messages
 from django.db import IntegrityError
-from django.shortcuts import render
-from django.shortcuts import redirect
+from django.shortcuts import render, redirect, reverse
 from django.contrib.auth import login, logout
-from .models import WelcomeRegister
+from .models import WelcomeRegister, Otp
 from .forms import RegisterForm, LoginForm, UserStudent
 from django.contrib.auth.models import User
 from .decorators import un_authenticated
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth import authenticate
+from .send_otp_sms import send_otp
+import uuid
+import random
 
 
 # Create your views here.
@@ -37,22 +39,29 @@ def user_register(request):
     if request.method == "POST":
         form = RegisterForm(request.POST)
         if form.is_valid():
-            try:
-                user_student, user = form.save()
-                login(request, user)
-                return redirect('home:home')
+            # check user is exists ?  if exist in database show error to user
+            phone_number = form.cleaned_data.get("phone_number")
+            if not User.objects.filter(username=phone_number).exists():
+                # Now create user and user student with overwrite save method and login user
+                try:
+                    user_student, user = form.save()
+                    login(request, user)
+                    return redirect('home:home')
 
-            except IntegrityError as e:
-                if 'phone_number' in str(e):
-                    form.add_error('phone_number', 'این شماره تلفن قبلاً ثبت شده است. لطفاً شماره دیگری انتخاب کنید.')
-                else:
+                except IntegrityError as e:
+                    if 'phone_number' in str(e):
+                        form.add_error('phone_number',
+                                       'این شماره تلفن قبلاً ثبت شده است. لطفاً شماره دیگری انتخاب کنید.')
+                    else:
+                        messages.error(request, f'خطایی رخ داده است: {str(e)}')
+                except Exception as e:
+                    username = form.cleaned_data.get('phone_number')
+                    if User.objects.filter(username=username).exists():
+                        User.objects.filter(username=username).delete()
+
                     messages.error(request, f'خطایی رخ داده است: {str(e)}')
-            except Exception as e:
-                username = form.cleaned_data.get('phone_number')
-                if User.objects.filter(username=username).exists():
-                    User.objects.filter(username=username).delete()
-
-                messages.error(request, f'خطایی رخ داده است: {str(e)}')
+            else:
+                form.add_error(None, 'این شماره قبلا ثبت شده است')
 
     return render(request, 'account/Signup.html', context={'welcome_text': welcome_text, 'form': form})
 
@@ -61,3 +70,13 @@ def user_register(request):
 def user_logout(request):
     logout(request)
     return redirect('home:home')
+
+
+def user_login_with_phone(request):
+    welcome_text = WelcomeRegister.objects.last()
+    if request.method == "POST":
+        token_otp = uuid.uuid4()
+        phone_number = request.GET.get('phone_number')
+        Otp.objects.create(phone_number=phone_number, token=token_otp, code=random.randint(1000, 9999))
+        return redirect(reverse('account:phoneRegister' + f'?token={token_otp}'))
+    return render(request, 'account/Login_with_phone.html', context={'welcome_text': welcome_text})
